+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [v2.1.2] 03-09-2016
+### Adds
+* New CLI arg to specify the html path
+* New CLI arg to enable the specification of a custom CSS, SCSS or HTML template.
+
+### Changes
+* HTML preview file name. Now, instead of "preview", it is named as "[fontname]-preview.html"
+
+### Fixes
+* Missing flags in the *--help* section (--cssFontsUrl, --cssDest and --sass)
+
 ## [v2.1.1] 02-09-2016
 ### Fixes
 * Missing documentation for *cssFontsUrl* and *cssDest* CLI args. (Issue Exictos-DCS/dcs-icon-font#2)